updates tables' structure

+++ backend/storages/DbStorage/migrations/1629737809810_initiate_users_table.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS users (
-    `id` int PRIMARY KEY auto_increment,
+    `id` bigint PRIMARY KEY auto_increment,
     `name` varchar(30) not null,
     `email` varchar(255) not null,
     `password` varchar(255) not null,
